Nova/requirements.txt
##########   Requirements nova automation ###################


APScheduler==3.7.0
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-apscheduler==0.6.0
django-filter==2.4.0
django-livereload-server==0.3.2
django-phonenumber-field==5.2.0
et-xmlfile==1.1.0
idna==2.10
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.20.3
oauthlib==3.1.0
openpyxl==3.0.7
pandas==1.2.4
phonenumbers==8.12.31
phonenumberslite==8.12.31
Pillow==8.2.0
pipe==1.6.0
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tk==0.1.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
